Severs the MFM parser from the overweight single MFM.hpp.
