+++ Storage/Disk/DiskImage/Formats/OricMFMDSK.cpp
@@ -10,6 +10,7 @@
 
 #include "../../Track/PCMTrack.hpp"
 #include "../../Encodings/MFM/MFM.hpp"
+#include "../../Encodings/MFM/Parser.hpp"
 
 using namespace Storage::Disk;
