+++ Storage/Disk/DiskImage/Formats/SSD.cpp
@@ -9,6 +9,7 @@
 #include "SSD.hpp"
 
 #include "../../Encodings/MFM/MFM.hpp"
+#include "../../Encodings/MFM/Parser.hpp"
 
 using namespace Storage::Disk;
